requirements.txt
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.1.0.post1
idna==3.4
mysql-connector-python==8.2.0
numpy==1.26.1
pandas==2.1.2
paramiko==3.3.1
protobuf==4.21.12
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
validate-email==1.3
